requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
numpy>=1.21.0
torch>=1.9.0
transformers>=4.11.0

# Speech Recognition
whisper>=1.0.0
soundfile>=0.10.3
librosa>=0.8.1

# Vector Database & RAG
faiss-cpu>=1.7.0
chromadb>=0.3.0
langchain>=0.0.150
llama-index>=0.6.0
sentence-transformers>=2.2.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pgvector>=0.1.0

# Audio Processing
pyaudio>=0.2.11
sounddevice>=0.4.4

# API & Web
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Monitoring & Logging
prometheus-client>=0.11.0
python-json-logger>=2.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.0 